pyproject.toml
[project]
name = "prosper-challenge"
version = "0.1.0"
description = "Template repository for clinic appointment scheduling voice agent challenge"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[webrtc,daily,silero,elevenlabs,openai,local-smart-turn-v3,runner]",
    "pipecat-ai-cli",
    "playwright",
    "psycopg[binary]>=3.3.3",
]

[dependency-groups]
dev = [
    "pyright>=1.1.404,<2",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "ruff>=0.12.11,<1",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]

[tool.pytest.ini_options]
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
]
pythonpath = ["."]

[tool.coverage.run]
branch = true
omit = [
    "src/bot.py",
    "src/integrations/healthie.py",
]

[tool.coverage.report]
skip_covered = true
